+++ cartridges/int_hipay_core/cartridge/scripts/init/hiPayServiceInit.js
@@ -176,10 +176,45 @@ function transaction() {
     return service;
 }
 
+/**
+ * Get available payment products.
+ * @returns {dw.svc.Service}
+ */
+function getAvailablePaymentProducts() {
+    var siteId = Site.getCurrent().getID();
+    var service = LocalServiceRegistry.createService('hipay.rest.available-payment-products.' + siteId, {
+        createRequest: function (svc, args) {
+            svc.setRequestMethod('GET');
+
+            // Set headers
+            svc.addHeader('Content-Type', 'application/json');
+            svc.addHeader('Cache-Control', 'no-cache');
+            svc.addHeader('Accept', 'application/json');
+
+            var credString = getCredentialsPrivate();
+            var base64Credentials = Encoding.toBase64(new Bytes(credString));
+            svc.addHeader('Authorization', 'Basic ' + base64Credentials);
+
+            svc.setURL(svc.getURL() + '?with_options=true&_format=json')
+
+            return '';
+        },
+        parseResponse: function (svc, response) {
+            return response;
+        },
+        filterLogMessage: function (msg) {
+            return msg;
+        }
+    });
+
+    return service;
+}
+
 module.exports = {
     order: order,
     hpayment: hpayment,
     maintenance: maintenance,
     apiData: apiData,
-    transaction: transaction
+    transaction: transaction,
+    getAvailablePaymentProducts: getAvailablePaymentProducts
 };