+++ CHANGELOG.md
@@ -1,27 +1,47 @@
-# Version 23.1.0
+# CHANGELOG
+
+## Version 25.4.0
+
+- **Add** : Added new payment means :
+  - Alma
+  - PayPal V2
+- **Add** : Updated one click payment
+- **Fix** : Fixed notification processing in correct order
+- **Fix** : Fixed iDeal payment method by removing `issuer_bank_id`
+
+## Version 23.1.0
+
 - Refactoring
 
-# Version 20.1.0
+## Version 20.1.0
+
 - Add DSP2 parameters
 
-# Version 2.1.0
+## Version 2.1.0
+
 - Update with latest module version
 
-# Version 2.0.0
+## Version 2.0.0
+
 - Migration to the Javascript controller.
 
-# Version 1.1.2
+## Version 1.1.2
+
 - Updates links to documentation.
 
-# Version 1.1.1
+## Version 1.1.1
+
 - Uses default config for a site if no specific one is found
 
-# Version 1.1.0
+## Version 1.1.0
+
 - Adds support for multi sites configuration
 
-# Version 1.0.1
+## Version 1.0.1
+
 - README file updates
 - License and contributing guidelines files added
 
-# Version 1.0.0
+## Version 1.0.0
+
 First production-ready version.